single version of calculateCountdown
